+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### ✨ Features
-- None
+- [QQ -> Discord] QQ 群的 `@全体成员` 现在会在 Discord 中转义为 `@everyone` 而不是仅展示文字
 
 ### 🐞 Bug Fixes
-- [QQ -> Discord] 修复了 QQ 消息中包含 `@everyone` 和 `@here` 时 Discord 会提及所有人的 bug [#8](https://github.com/Cola-Ace/koishi-plugin-bridge-discord-qq/issues/8)
+- None